Update Antimony docs
I'm trying to have a single source for all antimony docs, including this one.  In this commit, I round-tripped the file with pandoc from a .rst file to a .md file (the new 'origin of everything' format) and then back to .rst again.  This commit should thus be only formatting changes, with no content diff.