Round-trip antimony.rst through antimony.md with pandoc.
In cleaning up Antimony's documentation everywhere, I'm synchronizing everything to the .md form of this file.  But there are going to be some differences between this file and elsewhere, so as a first pass I'm just doing the round-trip reformatting of the file, so that other differences are easier to see.