Fix for issue 479: only set 'seed' once.
For this fix, I broke up the python construction of a simple task to the setup and the simulation, with the setup the part that sets the seed, and only the simulation the part that is executed inside the repeated task loop.

I think there might still be a problem with nested repeats, but with no working testing system, I didn't want to make this too complicated yet.