Resolves issue #484: allow KiSAO term 559 for initial time step.
KiSAO term 332 used to be used for 'initial time step', but that technically only applied to the RK45 algorithm.  The new 559 is more generic, can apply to any algorithm, and should be used instead.  Tellurium does not produce SED-ML, and there will be older files that use 332, so this fix allows both.