Use numpy >= 1.19.5
A Windows bug necessitated the use of numpy 1.19.3 and *not* 1.19.4.  The bug itself is being fixed in Windows, but in the meantime, numpy moved to a different system that doesn't fall prey to the bug, starting with 1.19.5.  This means that we can move our dependency to 1.19.5 and higher, which is more flexible than requiring 1.19.3 exactly (and allows us to incorporate other numpy fixes moving forward.)