Fix for *528:  new 'variable step size' KiSAO term.
There's a new KiSAO term that actually means 'the variable step size algorithm parameter', so add a new line that understands this.

Still have to deal with the issue of getting a uniform time course with a variable step size, which is an inherent contradiction, but this just adds it to the 'things we recognize' list, so we can deal with it properly elsewhere.